Broke logging code into more manageable sections
